Remove AGGREGATE attestation mode and Schnorr half-aggregation
Drop block-level signature aggregation (AGGREGATE=0x02) and the
BatchVerifier half-aggregation path. Individual Schnorr verification
per input is sufficient; half-aggregation added complexity without
meaningful savings at current transaction sizes.

- Delete aggregate.h/aggregate.cpp (AggregateProof, VerifyAggregateSpend,
  VerifyDeferredAttestation — all dead code)
- Remove BatchVerifier struct and Verify/FindFailure methods
- Remove batch-defer path in LadderSignatureChecker::CheckSchnorrSignature
- Remove AGGREGATE from RungAttestationMode enum and IsKnownAttestationMode
- Strip attestation field from coil RPC serialisation/parsing/help text
- Fix ComputeConditionsRoot → ComputeMLSCRoot in recursion tests
- Fix narrowing conversion warning in sighash.cpp
- Remove unused auto_tweaked variable in createtxmlsc
- Update aggregate attestation test to verify rejection
- Remove half_aggregation and batch_verifier test cases
- Mark transaction.h aggregated_sig as reserved (wire format compat)

MUSIG_THRESHOLD (0x0004) is unaffected — MuSig2/FROST key aggregation
is a distinct concept from block-level signature aggregation.

537 tests pass, 0 failures.

+++ src/rung/sighash.cpp
@@ -69,7 +69,7 @@ bool SignatureHashLadder(const PrecomputedTransactionData& cache,
     }
 
     const uint8_t base_hash_type = hash_type & 0x03; // strip APO/ACP flags
-    const uint8_t output_type = (hash_type == SIGHASH_DEFAULT || base_hash_type == 0) ? SIGHASH_ALL : base_hash_type;
+    const uint8_t output_type = (hash_type == SIGHASH_DEFAULT || base_hash_type == 0) ? static_cast<uint8_t>(SIGHASH_ALL) : base_hash_type;
     const uint8_t input_type = hash_type & SIGHASH_INPUT_MASK;
 
     HashWriter ss{HASHER_LADDERSIGHASH};
